HelpSidebar: Move component to HelpSidebar file
Its previous location, in ActivityContent.tsx, was a bit a fudge in
the first place, while developing the per-method IDE.
